requirements.txt
sphinx
sphinxnotes-any